Run ApexFile::Open in parallel during bootstrap
With the change in the SDV image (lunch target: `sdv_core_perf_arm64`), the 90th percentile ApexFile::Open time dropped by ~22% on Raspberry Pi 4B and ~35% on QC 8255, and the overall apexd-bootstrap stage decreased by ~14% on both Raspberry Pi 4B and QC 8255.

With the change in the AOSP image (lunch target:
`aosp_cf_x86_64_phone-trunk_staging-userdebug`), the 90th percentile ApexFile::Open time is reduced from 201ms to 184ms with an 8.5% improvement and the overall apexd-bootstrap process is reduced from 517ms to 464ms with a 10.3% improvement on Cuttlefish.

Open discussions carried over from aosp/3535045 as commented there:
- if using the new AddPreInstalledApexParallel for both bootstrap and
  non-bootstrap apexd stages
- if using the new ParseApexFiles (apex files are opened in parallel) to
  replace the OpenApexFiles in apexd.cpp

Test: atest ApexTestCases

Bug: 399312005
Change-Id: I72e2df42aaa79b7efeae269cce8fd893b012c429