Refactor apex-list-info.xml
It was bind-mounted to .default or .bootstrap apex-info-list.xml in each
mount namespace. It was because both namespaces share the same /apex
directory. Since each mount namespace has its own directory, the xml
file doesn't need to be bind-mounted.

In this change, emitting apex-info-list.xml code is deduped as well.

Bug: 365000206
Test: ApexTestCases
Test: manual test. $ adb shell ls -lZ /apex/apex-info-list.xml
Change-Id: I60d8253e867218d4b774f20913da6d53b07600eb