Fix calculateSizeForCompressedApex
Previously, it ignored an incoming compressed APEX when it has the same
version as the currently active decompressed version. Since same version
doesn't guarantee the same APEX, we can't ignore it.

Now, it's the only case we can ignore is that the incoming compressed
APEX is not selected due to the currently active "data" apex.

Fixed the logic accordingly.

Bug: 405231603
Test: ApexTestCases
Change-Id: Ia9475585380ea7a0f08dc68a08a1066d10846348